+++ scripts/build_and_run.sh
@@ -230,9 +230,9 @@ cat >"$INFO_PLIST" <<PLIST
   <key>CFBundlePackageType</key>
   <string>APPL</string>
   <key>CFBundleShortVersionString</key>
-  <string>1.6.1</string>
+  <string>1.6.9</string>
   <key>CFBundleVersion</key>
-  <string>9</string>
+  <string>17</string>
   <key>LSMinimumSystemVersion</key>
   <string>$MIN_SYSTEM_VERSION</string>
   <key>NSPrincipalClass</key>